test(consent): consolidate consents-tab tests, drop duplicate + flaky empty-state
Fold the non-duplicate tests from consents/encounterConsents.spec.ts into
the existing consent/consentCreation.spec.ts (search input, Add Consent
form defaults, encounter-actions command palette), removing the duplicate
'create with defaults' test already covered there. Drop the empty-state
test that asserted 'No Consents found' on the same shared fixture encounter
this file fills with consents (order-dependent/flaky). Delete the duplicate
consents/ folder to remove the consent vs consents confusion.
