--- tests/facility/patient/encounter/consents/encounterConsents.spec.ts
@@ -1,82 +1,0 @@
-import { expect, test } from "@playwright/test";
-import { clickTabOrMenuItem, openFixtureEncounter } from "tests/helper/ui";
-
-test.use({ storageState: "tests/.auth/user.json" });
-
-test.describe("Encounter Consents Tab", () => {
-  // The create test mutates the shared fixture encounter, so run this file's
-  // tests in declaration order (empty-state assertions before the create) in a
-  // single worker instead of in parallel.
-  test.describe.configure({ mode: "serial" });
-
-  test.beforeEach(async ({ page }) => {
-    await openFixtureEncounter(page);
-    await clickTabOrMenuItem(page, "Consents");
-    await expect(page).toHaveURL(/\/consents$/);
-  });
-
-  test("should display the consents tab with its empty state", async ({
-    page,
-  }) => {
-    await expect(
-      page.getByRole("tabpanel", { name: "Consents" }),
-    ).toBeVisible();
-    await expect(
-      page.getByRole("heading", { name: "No Consents found" }),
-    ).toBeVisible();
-  });
-
-  test("should have a search input for consents", async ({ page }) => {
-    await expect(
-      page.getByRole("textbox", { name: "Search existing consent" }),
-    ).toBeVisible();
-  });
-
-  test("should open the consent creation form", async ({ page }) => {
-    await page.getByRole("button", { name: "Add Consent" }).click();
-
-    const dialog = page.getByRole("dialog", { name: "Add Consent" });
-    await expect(dialog).toBeVisible();
-    await expect(dialog.getByRole("radio", { name: "Permit" })).toBeChecked();
-    await expect(
-      dialog.getByRole("combobox", { name: "Category" }),
-    ).toBeVisible();
-    await expect(dialog.getByRole("button", { name: "Save" })).toBeVisible();
-  });
-
-  test("should create a consent with the default fields", async ({ page }) => {
-    await page.getByRole("button", { name: "Add Consent" }).click();
-
-    const dialog = page.getByRole("dialog", { name: "Add Consent" });
-    await expect(dialog).toBeVisible();
-
-    // decision/category/status all have schema defaults (permit / treatment /
-    // active), so saving without changes submits a valid consent. The dialog
-    // closing below is what proves the save landed — a missing required field
-    // would keep it open and fail this test.
-    await dialog.getByRole("button", { name: "Save" }).click();
-
-    await expect(dialog).toBeHidden();
-    await expect(
-      page.getByRole("heading", { name: "No Consents found" }),
-    ).toBeHidden();
-  });
-
-  test("should open consents via the encounter actions command palette", async ({
-    page,
-  }) => {
-    await page.getByRole("tab", { name: "Overview" }).click();
-    await expect(page).toHaveURL(/\/updates$/);
-
-    await page
-      .getByRole("button", { name: /encounter actions/i })
-      .first()
-      .click();
-
-    const palette = page.getByRole("dialog", { name: "Command Palette" });
-    await expect(palette).toBeVisible();
-    await palette.getByRole("option", { name: /manage consents/i }).click();
-
-    await expect(page).toHaveURL(/\/consents$/);
-  });
-});
